requirements.txt
Django
djangorestframework
django-filter
django-oauth-toolkit
psycopg2
flask
pika
